rename loaders to skeleton
