+++ src/app/items/[id]/page.tsx
@@ -2,12 +2,12 @@
 import useDetailItem from "@/hooks/useDetailItem";
 import ItemDetail from "@/components/ItemDetail/ItemDetail";
 import styles from './page.module.scss'
-import ItemDetailLoader from "@/components/ItemDetailLoader/ItemDetailLoader";
+import ItemDetailLoader from "@/components/ItemDetailSkeleton/ItemDetailSkeleton";
 
 const ItemDetailPage = ({ params: { id } }: { params: { id: string } }) => {
     const { item, loading } = useDetailItem(id);
 
-    if (!item) return
+    if (!item && !loading) return
 
     return (
         <main className={styles.container}>
