+++ src/components/ItemDetailSkeleton/ItemDetailSkeleton.module.scss
